requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
greenlet>=3.0.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
asyncpg>=0.29.0
pgvector>=0.3.0
pypdf>=4.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-multipart>=0.0.6
sentence-transformers>=3.0.0
